added base for HDF5 tests
