+++ tests/test_grid_tools/test_hdf5.py
@@ -1,0 +1,9 @@
+import pytest
+
+
+class TestHDF5Creator:
+    pass
+
+
+class TestHDF5Interface:
+    pass
